Use OTel recommended names for attributes (#79)
* Use OTel recommended names for attributes

OpenTelemetry documents some conventional names to use for some pieces
of GraphQL server data. This change renames one attribute and adds two
more in line with OTel's recommendation, and also changes the name given
to the span that is started.

- Attribute "graphql.operation.name" was added
- Attribute "graphql.operation.type" was added

- BREAKING: "graphql.request.query" was renamed to "graphql.document".
- Span name was changed to `graphql.operation.type` +
  `graphql.operation.name` if those fields are present, else "GraphQL
  Operation"

See:
<https://opentelemetry.io/docs/reference/specification/trace/semantic_conventions/instrumentation/graphql/>

* Get operation name and type from blueprint

* Update changelog

* Document the trace_request_name and trace_request_type options

* Stringify the otel semantic convention attributes

Turns out tests break when we use atoms. Ideally we should look into
replacing all attributes with atoms instead, but for the sake of
simplicity this commit opts for using strings instead.

* Change behavior of the span_name config option

* Deprecate setting span_name

---------

Co-authored-by: mae.kasza <26093674+MaeIsBad@users.noreply.github.com>